Added hopscotch table implementation and rewrote the class table to use it.  This is done so that the class table and protocol tables can share the same code (currently there is no protocol table).
Note that concurrent resizing has not yet been implemented.  That means that there is a hard limit on the number of classes that can be loaded.  This is currently set to quite a small number for testing, to stress the hash table.  If you're experiencing problems as a result, then please let me know and I will increase it.



git-svn-id: svn://svn.gna.org/svn/gnustep/libs/libobjc2/trunk@29232 72102866-910b-0410-8b05-ffd578937521