Changed ptrdiff_t to int in property accessors.  This is for compatibility with some changes in clang that are required to prevent things from breaking on LP64 platforms.
It would be cleaner to use a ptrdiff_t here, but unfortunately we can't because the ivar_offset field in the runtime metadata is an int so ivar offsets beyond 4GB will not work on LP64 platforms.

This probably isn't a limitation.  If you have more than 4GB of ivars in one object, you've done something badly wrong and probably shouldn't be allowed to write code anymore.



git-svn-id: svn://svn.gna.org/svn/gnustep/libs/libobjc2/trunk@30024 72102866-910b-0410-8b05-ffd578937521