It turns out that Apple chose some confusing and non-intuitive semantics for catching Objective-C objects in C++ catch statements (i.e. they follow Objective-C semantics, not C++ semantics, irrespective of whether you use C++ or ObjC syntax).  We now default to Apple-compatible behaviour, but provide a function that allows users to select the sane semantics if they prefer.
Added a capability bit for the unified exception model, so code can require it.
This is not really required, since any code using it will link against the
ObjC++ personality function and will get a linker failure if it isn't supported.

Also enabled Objective-C++ stuff by default.  This adds a dependency on the C++
standard library (actually on libsupc++, but GNUstep Make wants to link against
libstd++ anyway), which is not ideal.  It can be disabled with:

$ gmake objectiver-cxx=no

I suggest that this is only done by people who know that they will never want
Objective-C++ support.



git-svn-id: svn://svn.gna.org/svn/gnustep/libs/libobjc2/trunk@32668 72102866-910b-0410-8b05-ffd578937521